docs: clarify AccessModal usage in routes
